Refactored code, controls now in separate file
